fix: avoid marking subtree as dynamic for inlined attributes
