+++ .changeset/spotty-sheep-fetch.md
@@ -1,0 +1,5 @@
+---
+'svelte': patch
+---
+
+fix: avoid marking subtree as dynamic for inlined attributes
